requirements.txt
psycopg2-binary>=2.9.0
numpy>=1.20.0
mmh3>=3.0.0
psutil>=5.9.0
